backend/pyproject.toml
[project]
name = "cleancar-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "celery>=5.4.0",
    "channels>=4.2.0",
    "channels-redis>=4.2.1",
    "daphne>=4.1.2",
    "django>=5.1.7",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.6",
    "python-decouple>=3.8",
    "uvicorn>=0.34.0",
    "websockets>=15.0.1",
]
